Sexual Activity in Risk Behavior Survey

Information on sexual activities from the (RBS)

Usage

data(sex)

Format

A tibble with 3,299 rows and 40 variables:

who

Type:	integer

Description:	Patient ID

txx_prt

Type:	factor (First/Reference level = 0)

Description:	Total sex partners:

0 = 0

1 = 1

2 = more than one

Levels:	0, 1, 2

txx_mprt

Type:	factor (First/Reference level = 0)

Description:	Total male sex partners:

0 = 0

1 = 1

2 = more than one

Levels:	0, 1, 2

txx_fprt

Type:	factor (First/Reference level = 0)

Description:	Total female sex partners:

0 = 0

1 = 1

2 = more than one

Levels:	0, 1, 2

msw_npt

Type:	factor (First/Reference level = 0)

Description:	Men who have sex with women (MSW), sex partners:

0 = 0

1 = 1

2 = more than one

Levels:	0, 1, 2

msw_frq

Type:	integer

Description:	MSW count of sex

msw_pxx

Type:	integer

Description:	MSW count of protected (with condom use) sex

msw_uxx

Type:	integer

Description:	MSW count of unprotected sex

msm_npt

Type:	factor (First/Reference level = 0)

Description:	Men have sex with men (MSM), sex partners:

0 = 0

1 = 1

2 = more than one

Levels:	0, 1, 2

msm_frq

Type:	integer

Description:	MSM count of sex

msm_pxx

Type:	integer

Description:	MSM count of protected (with condom use) sex

msm_uxx

Type:	integer

Description:	MSM count of unprotected sex

wsm_npt

Type:	factor (First/Reference level = 0)

Description:	Women have sex with men (WSM), sex partners:

0 = 0

1 = 1

2 = more than one

Levels:	0, 1, 2

wsm_frq

Type:	integer

Description:	WSM count of sex

wsm_pxx

Type:	integer

Description:	WSM count of protected (with condom use) sex

wsm_uxx

Type:	integer

Description:	WSM count of unprotected sex

txx_frq

Type:	integer

Description:	TOTAL count of sex

txx_pxx

Type:	integer

Description:	TOTAL count of protected (with condom use) sex

txx_uxx

Type:	integer

Description:	TOTAL count of unprotected sex

mvw_frq

Type:	integer

Description:	Men vaginal sex with women count of sex

maw_frq

Type:	integer

Description:	Men anal sex with women count of sex

mim_frq

Type:	integer

Description:	Men insertive sex with men count of sex

mrm_frq

Type:	integer

Description:	Men receptive sex with men count of sex

wvm_frq

Type:	integer

Description:	Women vaginal sex with men count of sex

wam_frq

Type:	integer

Description:	Women anal sex with men count of sex

mvw_pxx

Type:	integer

Description:	Men vaginal sex with women count of protected sex

maw_pxx

Type:	integer

Description:	Men anal sex with women count of protected sex

mim_pxx

Type:	integer

Description:	Men insertive sex with men count of protected sex

mrm_pxx

Type:	integer

Description:	Men receptive sex with men count of protected sex

wvm_pxx

Type:	integer

Description:	Women vaginal sex with men count of protected sex

wam_pxx

Type:	integer

Description:	Women anal sex with men count of protected sex

mvw_uxx

Type:	integer

Description:	Men vaginal sex with women count of unprotected sex

maw_uxx

Type:	integer

Description:	Men anal sex with women count of unprotected sex

mim_uxx

Type:	integer

Description:	Men insertive sex with men count of unprotected sex

mrm_uxx

Type:	integer

Description:	Men receptive sex with men count of unprotected sex

wvm_uxx

Type:	integer

Description:	Women vaginal sex with men count of unprotected sex

wam_uxx

Type:	integer

Description:	Women anal sex with men count of unprotected sex

t_p

Type:	integer

Description:	Total count of sex partners

t_fp

Type:	integer

Description:	Total count of female sex partners

t_mp

Type:	integer

Description:	Total count of male sex partners