Site Regrouped

This is study site (clinic/research site number) information. Large study sites were split into smaller groups and small study site were grouped into larger sites. The regrouped study sites are each approximately 100 subjects. This was done to protect the anonymity of the study sites.

Usage

data(site_masked)

Format

A tibble with 3,560 rows and 2 variables:

who

Type:	integer

Description:	Patient ID

site_masked

Type:	factor (First/Reference level = 270001)

Description:	Study Site Regrouped

Levels:	270001, 270002, 270003, 270004, 270005, 270006, 270007, 270008, 270009, 270010, 270011, 270012, 270013, 270014, 270015, 270016, 270017, 270018, 270019, 300001, 300002, 300003, 300004, 300005, 300006, 300007, 300008, 300009, 510001, 510002, 510003, 510004, 510005, 510006, 510007